// console-master/libs/data/eid/src/totp/totp-types.ts
//******************************************************************************
// Copyright 2021 BlackBerry. All Rights Reserved.

export interface TOTP {
  eecoid: string
  activated?: boolean
  created: string
  last_modified: string
}
// eslint-disable-next-line no-redeclare
export const TOTP = void 0